:recycle: refactor: Auth API Swagger 문서 추가 및 네이밍 변경 (#52)
